feat: add modicate controller
