+++ news-JS/src/components/controller/controller.ts
@@ -1,0 +1,43 @@
+import AppLoader from './appLoader';
+
+type Callback<TData> = (data: TData) => void;
+
+class AppController extends AppLoader {
+    getSources<TData>(callback: Callback<TData>) {
+        super.getResp(
+            {
+                endpoint: 'sources',
+            },
+            callback
+        );
+    }
+
+    getNews<TData>(e: Event, callback: Callback<TData>) {
+        let target = e.target as HTMLElement;
+        const newsContainer = e.currentTarget as HTMLElement;
+
+        while (target !== newsContainer) {
+            if (target.classList.contains('source__item')) {
+                const sourceId = target.getAttribute('data-source-id');
+                if (sourceId !== null) {
+                    if (newsContainer.getAttribute('data-source') !== sourceId) {
+                        newsContainer.setAttribute('data-source', sourceId);
+                        super.getResp(
+                            {
+                                endpoint: 'everything',
+                                options: {
+                                    sources: sourceId,
+                                },
+                            },
+                            callback
+                        );
+                    }
+                }
+                return;
+            }
+            target = target.parentNode as HTMLElement;
+        }
+    }
+}
+
+export default AppController;
